YVS/requirements.txt
youtube_transcript_api
streamlit
google_genai
python-dotenv
pathlib